Added common utility interface, which implements stl's random number generation & chrono utilities, also contains UpTime function for several platforms: Windows, Linux, Mac OS and FreeBSD. Fixed bitmap dimensions (int -> float) type in animinterface.h.

+++ src/animinterface.h
@@ -12,7 +12,7 @@ class AnimInterface {
 	ALLEGRO_VIDEO* mVideo = nullptr; 
 	CBEnchanted* mCBInstance = nullptr; 
 	ALLEGRO_BITMAP* mCurrentFrame = nullptr;
-	int mVideoWidth = 0, mVideoHeight = 0;
+	float mVideoWidth = 0, mVideoHeight = 0;
 	float mScaleX = 1.0f, mScaleY = 1.0f; 
 #endif 
 	public:
